refactor random image header to new file
