+++ client/components/Home.js
@@ -4,53 +4,20 @@ import HomeIntro from './HomeIntro/HomeIntro';
 import PortfolioOverview from './PortfolioOverview/PortfolioOverview';
 import TweetEmbed from 'react-tweet-embed'
 import Footer from './Footer/Footer';
-import {headerLayouts} from '../data/header-layouts';
+const randomHeader = require('./Header/randomHeader');
 import CustomProperties from 'react-custom-properties';
-
-const importHeaderImages = (files) => {
-  let images = {};
-  files.keys().map((item, index) => { images[item.replace('./', '').replace(/\.[^/.]+$/, '')] = files(item); });
-  return images;
-};
-const allHeaderImages = importHeaderImages(require.context('../assets/images/headers', false, /\.(png|jpe?g)$/));
 
 class Home extends React.Component {
   constructor(props) {
     super(props);
-    this.randomheaderRange = this.randomheaderRange.bind(this);
-    this.randomizeHeader = this.randomizeHeader.bind(this);
-    this.randomizeHeaderObject = this.randomizeHeaderObject.bind(this);
 
     this.state = {
-      'randomPhoto': this.randomheaderRange(),
       'randomPhotoObject': {}
     };
   }
 
-  componentDidMount() {
-    this.randomizeHeaderObject();
-  }
-
-  randomheaderRange(min = 1, max = 10) {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  }
-
-  randomizeHeader() {
-    const random = this.randomheaderRange();
-    this.setState({'randomPhoto': random});
-  }
-
-  randomizeHeaderObject() {
-    const randomPhotoObject = {...this.state.randomPhotoObject};
-    const random = this.randomheaderRange();
-    const headerLayout = headerLayouts[random];
-    const headerImage = allHeaderImages[headerLayout.image];
-    randomPhotoObject.image = headerImage;
-    randomPhotoObject.color = headerLayout.color;
-    randomPhotoObject.backgroundColor = headerLayout.backgroundColor;
-    randomPhotoObject.blend = headerLayout.blend;
-    randomPhotoObject.linkColor = headerLayout.linkColor;
-    this.setState({randomPhotoObject});
+  componentWillMount() {
+    randomHeader.randomizeHeader(this);
   }
 
   render() {
@@ -68,7 +35,7 @@ class Home extends React.Component {
           global
           properties={headerStyles}
         />
-        <Header randomizeHeader={this.randomizeHeaderObject}/>
+        <Header randomizeHeader={randomHeader.randomizeHeader}/>
         <div className="page--home page-content">
           <h1 className="page-title">
             Hi. <span className="my-name">I’m <span className="no-break">Mason Wendell.</span></span>
